Added bin object

+++ src/main/java/entity/SelectionCircle.java
@@ -3,6 +3,7 @@ package entity;
 import core.CollisionBox;
 import core.Position;
 import core.Size;
+import entity.scenery.Bin;
 import entity.scenery.Scenery;
 import state.State;
 import graphics.ImageUtils;
@@ -21,7 +22,7 @@ public class SelectionCircle  extends GameObject {
 
     public SelectionCircle(Scenery scenery) {
         this.color = new Color(255, 255, 255, 150);
-        this.size = new Size(scenery.getSprite().getWidth(null), (int) scenery.getSprite().getHeight(null) / 6);
+        this.size = new Size(scenery.getSprite().getWidth(null), scenery.getSprite().getHeight(null) / 6);
         renderOffset = scenery.getSelectionCircleOffset();
         collisionBoxOffset = renderOffset;
         renderOrder = 4;
